voice/requirements.txt
pyttsx3
speechrecognition
google-generativeai
python-dotenv


